Fixed infinite API call bug
