+++ FlavorSaver/Models/Search.swift
@@ -13,19 +13,11 @@ class Search : ObservableObject{
     private var selectedIngredients : [String] = []
     @Published private var searchResults : RecipesMetaData = RecipesMetaData(offset: 0, numberOfRecipes: 0, totalRecipes: 0, recipes: [])
     @Published private var listOfRecipes : [(String,[Recipe])] = []
-    @Published private var recommendedRecipes : [Recipe] = []
     
     private var ingredientFinder : Ingredients = Ingredients()
     private var hasChanged : Bool = true
     private var apiManager = APIManager()
-    
-    init(){
-        Task(priority: .high){
-            await executeRandomSearch()
-        }
-    }
-    
-    
+        
     /// Adds an ingredient to the search request. No ingredients are added if the ingredient is already in the list
     /// - Parameter ingredient: the ingredient to add
     func addIngredient(_ ingredient : String) {
@@ -51,12 +43,6 @@ class Search : ObservableObject{
         return selectedIngredients
     }
     
-    /// Returns the current list of recommended recipes
-    /// - Returns: list of recipes
-    func getRecommendedRecipes() -> [Recipe]{
-        return recommendedRecipes
-    }
-    
     /// Returns a list of all recipes returned from the search parameter
     /// - Returns: list of recipes
     func getRecipes() -> [Recipe]{
@@ -67,25 +53,6 @@ class Search : ObservableObject{
     /// - Returns: list of recipes
     func getRecipesWithTags() -> [(String,[Recipe])]{
         return listOfRecipes
-    }
-    
-    /// Executes a request to update the random recommended recipes
-    func executeRandomSearch() async {
-        let urlRequest = "\(apiManager.randomSearchParams)"
-        do{
-            let request = try await apiManager.sendAPIRequest(urlRequest, RandomRecipes.self)
-            
-            DispatchQueue.main.async{
-                self.recommendedRecipes = request.recipes
-            }
-        }catch{
-            if (apiManager.apiVersion == "SPOON_API"){
-                apiManager.apiVersion = "SPOON_API2"
-                await executeRandomSearch()
-            }else{
-                print("Error retrieving the recipes in Search.executeRandomSearch(). Error: \(error)")
-            }
-        }
     }
     
     /// Executes the search request and updates the @Published listOfRecipes retrieved from getRecipesWithTags or getRecipes
